requirements.txt
tk
pyserial-asyncio